+++ lib/kucoin_ruby/exceptions/base.rb
@@ -1,0 +1,6 @@
+module KucoinRuby
+  module Exceptions
+    class Base < StandardError
+    end
+  end
+end