+++ lib/kucoin_ruby/util.rb
@@ -4,8 +4,8 @@ module KucoinRuby
       (Time.now.to_f * 1000).to_i.to_s
     end
 
-    def self.build_message(endpoint, nonce, queryString = '')
-      "#{endpoint}/#{nonce}/#{queryString.to_param}"
+    def self.build_message(endpoint, nonce, query_string = '')
+      "#{endpoint}/#{nonce}/#{query_string}"
     end
 
     def self.sign(string)
@@ -13,15 +13,23 @@ module KucoinRuby
       sign_hmac(base)
     end
 
-    def self.sign_message(endpoint, queryString = '')
+    def self.sign_message(endpoint, query_string = nil)
       nc = nonce
-      [nc, sign(build_message(endpoint, nc, queryString.to_param))]
+      query_string = compose_query_string(query_string)
+      [nc, sign(build_message(endpoint, nc, query_string))]
     end
 
     private
 
     def self.sign_hmac(string)
-      OpenSSL::HMAC.hexdigest(OpenSSL::Digest.new('sha256'), KucoinRuby::NetConfig.secret, string)
+      OpenSSL::HMAC.hexdigest(OpenSSL::Digest.new('sha256'), KucoinRuby::Net.secret, string)
+    end
+
+    def self.compose_query_string(query_string)
+      return query_string unless query_string.is_a? Hash
+      uri = Addressable::URI.new
+      uri.query_values = query_string
+      uri.query
     end
   end
 end